+++ test/functional-tests/node-transformers/converting-transformers/object-pattern-properties-transformer/fixtures/static-block-destructuring-assignment.js
@@ -1,0 +1,11 @@
+(function() {
+    class Foo {
+        static result;
+        static {
+            let x;
+            ({ x } = { x: 42 });
+            Foo.result = x;
+        }
+    }
+    return Foo.result;
+})()