+++ test/functional-tests/node-transformers/converting-transformers/object-pattern-properties-transformer/fixtures/static-block-destructuring-declaration.js
@@ -1,0 +1,10 @@
+(function() {
+    class Foo {
+        static result;
+        static {
+            let { x } = { x: 42 };
+            Foo.result = x;
+        }
+    }
+    return Foo.result;
+})()