+++ test/functional-tests/node-transformers/rename-identifiers-transformers/scope-identifiers-transformer/class-declaration/ClassDeclaration.spec.ts
@@ -805,5 +805,6 @@ describe('ScopeIdentifiersTransformer ClassDeclaration identifiers', () => {
                 assert.match(obfuscatedCode, defaultExportRegExp);
             });
         });
+
     });
 });